% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.costandusagereportservice_operations.R
\name{delete_report_definition}
\alias{delete_report_definition}
\title{Delete a specified report definition}
\usage{
delete_report_definition(ReportName = NULL)
}
\arguments{
\item{ReportName}{}
}
\description{
Delete a specified report definition
}
\examples{

}
