% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.costandusagereportservice_operations.R
\name{describe_report_definitions}
\alias{describe_report_definitions}
\title{Describe a list of report definitions owned by the account}
\usage{
describe_report_definitions(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Describe a list of report definitions owned by the account
}
\examples{

}
