% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.costandusagereportservice_package.R
\docType{package}
\name{paws.costandusagereportservice-package}
\alias{paws.costandusagereportservice}
\alias{paws.costandusagereportservice-package}
\title{paws.costandusagereportservice: AWS Cost and Usage Report Service}
\description{
All public APIs for AWS Cost and Usage Report service
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
