% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.costandusagereportservice_operations.R
\name{put_report_definition}
\alias{put_report_definition}
\title{Create a new report definition}
\usage{
put_report_definition(ReportDefinition)
}
\arguments{
\item{ReportDefinition}{}
}
\description{
Create a new report definition
}
\examples{

}
