% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.costandusagereportservice_operations.R
\name{put_report_definition}
\alias{put_report_definition}
\title{Create a new report definition}
\usage{
put_report_definition(
  ReportDefinition = list(
    ReportName = "string",
    TimeUnit = "HOURLY"|"DAILY",
    Format = "textORcsv",
    Compression = "ZIP"|"GZIP",
    AdditionalSchemaElements = list(
      "RESOURCES"
    ),
    S3Bucket = "string",
    S3Prefix = "string",
    S3Region = "us-east-1"|"us-west-1"|"us-west-2"|"eu-central-1"|"eu-west-1"|"ap-southeast-1"|"ap-southeast-2"|"ap-northeast-1",
    AdditionalArtifacts = list(
      "REDSHIFT"|"QUICKSIGHT"
    )
  )
)
}
\arguments{
\item{ReportDefinition}{[required]}
}
\description{
Create a new report definition
}
\examples{

}
